pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelsteg"
version = "0.1.0"
description = "Hide and reveal data in PNG images using LSB steganography"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = ["Pillow>=10.0"]

[project.scripts]
steg = "pixelsteg.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pixelsteg*"]

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
